git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/split_over_and_underflow@37780 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/automaton/BuechiConverterUtils.java
@@ -203,7 +203,7 @@ public class BuechiConverterUtils {
             new ImmutableList.Builder<>();
 
         StoredState initBuchiState =
-            storedAutomaton.getStoredState(Iterables.getOnlyElement(initStateList).intValue());
+            storedAutomaton.getStoredState(Iterables.getOnlyElement(initStateList));
         String initBuechiStateName = getStateName(initBuchiState);
 
         String initStateName = null;
@@ -220,7 +220,7 @@ public class BuechiConverterUtils {
           List<AutomatonTransition> transitionList = new ArrayList<>();
 
           for (StoredEdgeWithLabel edge : storedAutomaton.getEdgesWithLabel(i)) {
-            int successorStateId = Iterables.getOnlyElement(edge.getConjSuccessors()).intValue();
+            int successorStateId = Iterables.getOnlyElement(edge.getConjSuccessors());
             String successorName = getStateName(storedAutomaton.getStoredState(successorStateId));
 
             transitionList.addAll(getTransitions(edge.getLabelExpr(), successorName));
